==> notebooks/churn_project/02_Preprocesing.ipynb <==
import pandas as pd
import joblib
from sklearn.preprocessing import LabelEncoder
from sklearn.preprocessing import StandardScaler

pd.set_option("display.max_rows", 10)
customer_churn_reviews = pd.read_csv("../../data/churn_data.csv", index_col = 0)

# 1. 결측치 확인하기

# 각 컬럼별 결측치 개수 확인
customer_churn_reviews.isnull().sum()

#확인 결과 결측치의 개수가 0이므로 결측치 처리는 건너뛴다.

#머신러닝을 쓸 예정이기에, 문자열(범주형 변수)는 숫자형으로 변한이 필요.

#범주형 변수 인코딩 방법에는 2가지가 있음.
#1. Label Encoding : 각 볌주를 정수 값으로 변환, gender -> Male(0), Female(1) 간단하나, 숫자로 변환하면 모델이 순서를 오해할 수 있음
#2. One-Hot Encoding : 컬럼을 범주별로 새로 만들어서 0/1로 표시, InternetService -> DSL, Fiber optic, NO 3개의 컬럼 생성, 단점 : 컬럼 수가 늘어난다.

#성별 같은 단순한 이진 변수 -> Label Encoding
#3개 이상 범주가 있는 컬럼(PaymentMethod, Contract 등) -> One-Hot Encoding

#따라서 컬럼 별로 이진 변수인지, 3개 이상 범주형 변수인지 나눈다.

#범주형 컬럼 찾기
categorical_cols = customer_churn_reviews.select_dtypes(include=['object']).columns

#이진 변수 / 다중 범주 변수 구분
binary_vars = []
multi_vars = []

for col in categorical_cols:
    unique_vals = customer_churn_reviews[col].nunique() #고유값 개수
    if unique_vals == 2:
        binary_vars.append(col) #값이 2개 -> 이진 변수
    elif unique_vals > 2:
        multi_vars.append(col) #값이 3개 이상 -> 다중 변수

print("이진 변수(Binary Variables):", binary_vars) #Label Encoding 할 변수
print("다중 범주 변수(Multi-class Variables):", multi_vars) #One-Hot Encoding 할 변수

# 1. 이진 변수 Label Encoding
binary_vars = ['gender', 'Partner', 'Dependents', 'PhoneService', 'PaperlessBilling']
le = LabelEncoder()
for col in binary_vars:
    customer_churn_reviews[col] = le.fit_transform(customer_churn_reviews[col])

# Churn(타겟)도 Yes=1, No=0으로 인코딩
customer_churn_reviews['Churn'] = customer_churn_reviews['Churn'].map({'Yes': 1, 'No': 0})

# 2. 다중 범주형 변수 One-Hot Encoding
multi_vars = ['MultipleLines', 'InternetService', 'OnlineSecurity', 'OnlineBackup',
              'DeviceProtection', 'TechSupport', 'StreamingTV', 'StreamingMovies',
              'Contract', 'PaymentMethod']

customer_churn_reviews = pd.get_dummies(customer_churn_reviews, columns=multi_vars)

print(customer_churn_reviews.head())

#get_dummies 코드가 있을 때 dtype 메소드를 실행시키지 말 것. 
#Jupyter 특성상 같은 셀 안에서 호출되면 원본 컬럼을 못 찾아서 KeyError 발생

customer_churn_reviews.dtypes

# TotalCharges가 Object로 읽혀서, 숫자로 변환.
# 숫자로 변환 (문자열 → NaN 변환 허용)
customer_churn_reviews['TotalCharges'] = pd.to_numeric(customer_churn_reviews['TotalCharges'], errors='coerce')

# NaN이 생겼다면 중앙값으로 대체
customer_churn_reviews['TotalCharges'] = customer_churn_reviews['TotalCharges'].fillna(customer_churn_reviews['TotalCharges'].median())

#수치형 변수만 추출
numeric_cols = customer_churn_reviews.select_dtypes(include=['int64', 'float64']).columns
print(numeric_cols)

#연속형 변수를 직접 확인
for col in numeric_cols:
    print(col, customer_churn_reviews[col].nunique())

#실행 결과 tenure, MontlyCharges (TotalCharges가 나와야 하는데 Object로 읽히는 오류가 있음)
#따라서 [26]에서 변환해주는 과정을 거침.

scaler = StandardScaler()
num_vars = ['tenure', 'MonthlyCharges', 'TotalCharges']
customer_churn_reviews[num_vars] = scaler.fit_transform(customer_churn_reviews[num_vars])

#스케일링이 제대로 됐는지 확인
print(customer_churn_reviews[num_vars].describe())

# -----------------------
# 전처리 완료 데이터 저장
# -----------------------
customer_churn_reviews.to_csv("../../data/customer_churn_preprocessed.csv", index=False)
print("✅ 전처리된 데이터 저장 완료: data/customer_churn_preprocessed.csv")

#04_evalutation에서 파일 불러올 때 오류가 나서 원인 해결을 위해 feature 목록도 같이 저장
joblib.dump(list(customer_churn_reviews.drop("Churn", axis=1).columns),
            "../../results/feature_names.pkl")
print("✅ feature 데이터 저장 완료: results/feature_names.csv")


print(customer_churn_reviews.columns.tolist())
