==> notebooks/churn_project/04_evaluation_report.ipynb <==
import joblib
import pandas as pd
from sklearn.model_selection import train_test_split



#저장된 모델, feature 불러오기
lg_loaded = joblib.load("../../results/final_lg_model.pkl")
feature_names = joblib.load("../../results/feature_names.pkl")

#전처리된 데이터 불러오기
customer_churn_reviews = pd.read_csv("../../data/customer_churn_preprocessed.csv", index_col = 0)
X = df.drop('Churn', axis=1)
y = df['Churn']

#feature 순서 맞추기
X = X[feature_names]

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.3, random_state=42, stratify=y
)

#예측 확률 불러오기 (ROC Curve 등 시각화에 필요)
y_proba = lg_loaded.predict_proba(X_test)[:,1]
y_pred = (y_proba >= 0.25).astype(int) #threshold 조정 가능

df = pd.read_csv("../../data/customer_churn_preprocessed.csv")
print(df.head())
print(df.dtypes)
